Feature/shoot (#8)
* First pass on shooting

* Removed rotations for now.

* Updated Update method to use double instead of floats, Updated the shoot method to move depending on the ship's orientation.

* Made HandleInput public, added collision checks for asteroids against player and against bullets. Added bullet lifespan calculation.

* Fixed some bugs related to bullet collision, added some extra getters to the entity class. Added new respaning state. Added check for respawning when shooting.

* Fixed x86 builds.

* Added missing external deps

+++ BoilerPlate/Asteroid.cpp
@@ -56,11 +56,11 @@ namespace Asteroids
 			ApplyRandomImpulse();
 		}
 
-		void Asteroid::Update(float deltaTime)
+		void Asteroid::Update(double deltaTime)
 		{
 			// Rotate the asteroid.
 			//
-			float angle = m_transforms->GetAngleInDegrees() + ROTATION_SPEED * deltaTime;
+			float angle = m_transforms->GetAngleInDegrees() + ROTATION_SPEED * static_cast<float>(deltaTime);
 			m_transforms->RotateInDegrees(angle);
 
 			Entity::Update(deltaTime);
